Refactor Becoming architecture: Separate coordination from execution
Revolutionary architectural change separating the Becoming coordinator
from the Being executor. This achieves pure ontological programming
with beautiful philosophical method naming.

Key changes:
- Create Being class to handle transformation execution
- Add willBe() method: "What will you become?"
- Add metamorphose() method: Complete transformation
- Update Becoming to pure coordination role
- Update BecomingArgumentsInterface: __invoke → be()

Philosophy: "Life as continuous becoming: Being reveals its becoming, then becomes it"
Template pattern established for all future ontological interactions.
