--- v0/src/FakeProcessedData.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Be\Framework;
-
-use Ray\InputQuery\Attribute\Input;
-
-/**
- * Fake processed data class for testing
- */
-final class FakeProcessedData
-{
-    public function __construct(
-        #[Input]
-        public readonly string $data,
-    ) {
-    }
-}
